Done setup

+++ backend/controllers/product.controller.js
@@ -52,11 +52,18 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid Product Id" });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ success: true, message: "Product deleted" });
   } catch (error) {
     console.log("Error in deleting product: ", error.message);
-    res.status(404).json({ success: false, message: "Product Notfound" });
+    res.status(500).json({ success: false, message: "Server error" });
   }
 };
